Move server.Config to config.Config
