+++ controller/application_controller.go
@@ -1,19 +1,19 @@
 package controller
 
 import (
+	"github.com/dtan4/paus-frontend/config"
 	"github.com/dtan4/paus-frontend/model/user"
-	"github.com/dtan4/paus-frontend/server"
 	"github.com/dtan4/paus-frontend/store"
 	"github.com/gin-gonic/contrib/sessions"
 	"github.com/gin-gonic/gin"
 )
 
 type ApplicationController struct {
-	config *server.Config
+	config *config.Config
 	etcd   *store.Etcd
 }
 
-func NewApplicationController(config *server.Config, etcd *store.Etcd) *ApplicationController {
+func NewApplicationController(config *config.Config, etcd *store.Etcd) *ApplicationController {
 	return &ApplicationController{
 		config: config,
 		etcd:   etcd,
